new version! of printf

+++ libft/ft_strtrim.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strtrim.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hladeiro <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: hladeiro <hladeiro@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/04/16 14:15:35 by hladeiro          #+#    #+#             */
-/*   Updated: 2024/04/16 16:02:38 by hladeiro         ###   ########.fr       */
+/*   Created: 2024/04/18 19:18:48 by hladeiro          #+#    #+#             */
+/*   Updated: 2024/04/18 19:18:49 by hladeiro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,12 +26,12 @@ static int	trim_words(const char *set, char c)
 	return (0);
 }
 
-static char	*trim_string(const char *s1, int start, int len)
+static char	*trim_string(const char *s1, size_t start, size_t len)
 {
 	char	*str;
-	int		i;
+	size_t	i;
 
-	if (start >= (int)ft_strlen(s1))
+	if (start >= ft_strlen(s1))
 		return (ft_strdup(""));
 	str = ft_calloc(len + 1, sizeof(char));
 	if (!str)
@@ -47,15 +47,15 @@ static char	*trim_string(const char *s1, int start, int len)
 
 char	*ft_strtrim(const char *s1, const char *set)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	if (ft_strlen(s1) == 0)
 		return (ft_strdup(""));
 	if (ft_strlen(set) == 0)
 		return (ft_strdup(s1));
 	i = 0;
-	j = (int) ft_strlen(s1) - 1;
+	j = ft_strlen(s1) - 1;
 	while (trim_words(set, s1[i]))
 		i++;
 	while (trim_words(set, s1[j]))
